display rstack too
but I probably won't end up displaying globals...? because there isn't a
"natural" way of displaying them, and they can easily be "printf-debugged" if
needed?

also now sending objects as objects instead of object-like arrays. I'm kinda
depending on JSON.stringify(a)===JSON.stringify(b) as a deep_equals(a,b) kinda
thing, which relies on key ordering being stable in JS, which it is. it's still
subtly unnerving, and feels like something that might break.