"break before"
I've surrendered to "break before op" instead of "break after op". there are a
lot of subtle problems with "break after", maybe even for the user...  like
what does "break after" mean when the cursor is on "if"?

also... there can only be one user breakpoint now, and it is no longer handled
as a program-edit, but rather as an arg/pos. thus, tests that have multiple
breakpoints sets each breakpoint individually now.

the source lookup code has also improved; it now decides which "word span" the
cursor is in before trying to find the op. this is better than the previous
code that had trouble figuring out which word the cursor is in sometimes.